Query Causes with artists
Preload artists in Causes Query

- Move the set up of the context object to schema.ex
- Use Absinthe dataloader to resolve and preload artists in CauseType
- Update module name for the types to SingForNeedsWeb for better namespacing
