+++ .iex.exs
@@ -1,5 +1,7 @@
+require Ecto.Query
 alias SingForNeeds.Artists
 alias SingForNeeds.Artists.Artist
 alias SingForNeeds.Causes
 alias SingForNeeds.Causes.Cause
 alias SingForNeeds.Repo
+alias Ecto.Query
