+++ lib/sing_for_needs_web/resolvers/performances.ex
@@ -1,4 +1,4 @@
-defmodule SingForNeeds.Resolvers.Performance do
+defmodule SingForNeedsWeb.Resolvers.Performance do
   @moduledoc """
   Resolves for performances
   """
