+++ lib/sing_for_needs_web/schema/schema.ex
@@ -5,7 +5,24 @@ defmodule SingForNeedsWeb.Schema.Schema do
   use Absinthe.Schema
   import_types(Absinthe.Type.Custom)
   import_types(SingForNeedsWeb.Schema.{ArtistTypes, CauseTypes, PerformanceTypes})
-  alias SingForNeeds.Resolvers.{Artist, Cause, Performance}
+  alias SingForNeedsWeb.Resolvers.{Artist, Cause, Performance}
+
+  def dataloader do
+    alias SingForNeeds.{Artists, Causes}
+
+    loader =
+      Dataloader.new()
+      |> Dataloader.add_source(Causes, Causes.datasource())
+      |> Dataloader.add_source(Artists, Artists.datasource())
+  end
+
+  def context(ctx) do
+    Map.put(ctx, :loader, dataloader())
+  end
+
+  def plugins do
+    [Absinthe.Middleware.Dataloader] ++ Absinthe.Plugin.defaults()
+  end
 
   query do
     @desc "get list of artists"
